Explain 'pcsc-shared' in nitrokey3/linux/fedora-gnupg-configuration
A common issue with scdaemon is that it wants exclusive access to pcscd,
causing gpg to fail without explanation when a browser is open.
This doesn't seem to resolve until the browser is closed, making gpg
essentially unusable. The `pcsc-shared` option fixes this.

It isn't default, because it can cause problems when other clients
concurrently write to the card. But as far as I understand, it's
unlikely that a user has one client writing to the card while also
making another operation with gpg/scdaemon.

References:
* https://wiki.archlinux.org/title/GnuPG#Shared_access_with_pcscd
* https://www.gnupg.org/documentation/manuals/gnupg/Scdaemon-Options.html
* gpg 2.2.28 release: https://lists.gnupg.org/pipermail/gnupg-announce/2021q2/000460.html