feat: add SELinux contexts, mention relevance for other Nitrokey Apps
I added the mention that these rules also apply for all other nitrokey apps.

Additionally added a command to fix the files to carry the correct SELinux labels, as those are

```
unconfined_u:object_r:user_home_t:s0 41-nitrokey.rules
```

by default. This can be an issue and should be implemented in as many locations as possible.

`mv -Z` does not work either, as `mv` is SELinux aware but that argument makes it keep the default context unconfined.

A systemwide relabelling may be an alternative solution, which should not cause any breakages and would fixup incorrect contexts in many other areas.